feat: restructure the project into separate binaries and library
