+++ Cargo.toml
@@ -17,6 +17,18 @@ keywords = ["bioinformatics", "amino-acids"]
 license = "MIT OR Apache-2.0"
 readme = "README.md"
 
+[lib]
+crate-type = ["dylib", "lib", "staticlib"]
+path = "src/aaprop_lib/lib.rs"
+
+[[bin]]
+name = "aaprop-server"
+path = "src/aaprop_server/main.rs"
+
+[[bin]]
+name = "aaprop"
+path = "src/aaprop_shuttle/main.rs"
+
 [features]
 shuttle = ["tracing"]
 tracing = []
@@ -27,8 +39,8 @@ server = []
 axum = "0.7.5"
 serde = { version = "1.0.197", features = ["derive", "rc"] }
 serde_json = "1.0.115"
-shuttle-axum = "0.41.0"
-shuttle-runtime = "0.41.0"
+shuttle-axum = "0.42.0"
+shuttle-runtime = "0.42.0"
 strum = { version = "0.26.2", features = ["derive", "strum_macros"] }
 strum_macros = "0.26.2"
 tokio = "1.36.0"
@@ -36,6 +48,9 @@ tracing = { version = "0.1.38", features = ["async-await"] }
 
 [dev-dependencies]
 rstest = "0.18.2"
+
+[lints.rust]
+unsafe_code = "forbid" # Ensure that the crate does not use unsafe code
 
 [lints.rustdoc]
 broken_intra_doc_links = "deny" # Ensure that all intra-doc links point to valid items
