+++ src/aaprop_lib/amino_acid_data.json
@@ -1,0 +1,242 @@
+[
+    {
+        "abbreviation": "A",
+        "codons": [
+            "GCT",
+            "GCC",
+            "GCA",
+            "GCG"
+        ],
+        "molecular_weight": 89.09,
+        "name": "Alanine",
+        "short_name": "Ala",
+        "side_chain": "Nonpolar"
+    },
+    {
+        "abbreviation": "R",
+        "codons": [
+            "CGT",
+            "CGC",
+            "CGA",
+            "CGG",
+            "AGA",
+            "AGG"
+        ],
+        "molecular_weight": 174.2,
+        "name": "Arginine",
+        "short_name": "Arg",
+        "side_chain": "Basic"
+    },
+    {
+        "abbreviation": "N",
+        "codons": [
+            "AAT",
+            "AAC"
+        ],
+        "molecular_weight": 132.12,
+        "name": "Asparagine",
+        "short_name": "Asn",
+        "side_chain": "Polar"
+    },
+    {
+        "abbreviation": "D",
+        "codons": [
+            "GAU",
+            "GAC"
+        ],
+        "molecular_weight": 133.1,
+        "name": "Aspartic Acid",
+        "short_name": "Asp",
+        "side_chain": "Acidic"
+    },
+    {
+        "abbreviation": "C",
+        "codons": [
+            "TGT",
+            "TGC"
+        ],
+        "molecular_weight": 121.15,
+        "name": "Cysteine",
+        "short_name": "Cys",
+        "side_chain": "Polar"
+    },
+    {
+        "abbreviation": "E",
+        "codons": [
+            "GAA",
+            "GAG"
+        ],
+        "molecular_weight": 147.13,
+        "name": "Glutamic Acid",
+        "short_name": "Glu",
+        "side_chain": "Acidic"
+    },
+    {
+        "abbreviation": "Q",
+        "codons": [
+            "CAA",
+            "CAG"
+        ],
+        "molecular_weight": 146.15,
+        "name": "Glutamine",
+        "short_name": "Gln",
+        "side_chain": "Polar"
+    },
+    {
+        "abbreviation": "G",
+        "codons": [
+            "GGT",
+            "GGC",
+            "GGA",
+            "GGG"
+        ],
+        "molecular_weight": 75.07,
+        "name": "Glycine",
+        "short_name": "Gly",
+        "side_chain": "Nonpolar"
+    },
+    {
+        "abbreviation": "H",
+        "codons": [
+            "CAT",
+            "CAC"
+        ],
+        "molecular_weight": 155.16,
+        "name": "Histidine",
+        "short_name": "His",
+        "side_chain": "Positive"
+    },
+    {
+        "abbreviation": "I",
+        "codons": [
+            "ATT",
+            "ATC",
+            "ATA"
+        ],
+        "molecular_weight": 131.18,
+        "name": "Isoleucine",
+        "short_name": "Ile",
+        "side_chain": "Nonpolar"
+    },
+    {
+        "abbreviation": "L",
+        "codons": [
+            "TTA",
+            "CTT",
+            "CTC",
+            "CTG",
+            "CTA"
+        ],
+        "molecular_weight": 131.18,
+        "name": "Leucine",
+        "short_name": "Leu",
+        "side_chain": "Nonpolar"
+    },
+    {
+        "abbreviation": "K",
+        "codons": [
+            "AAA",
+            "AAG"
+        ],
+        "molecular_weight": 146.19,
+        "name": "Lysine",
+        "short_name": "Lys",
+        "side_chain": "Basic"
+    },
+    {
+        "abbreviation": "M",
+        "codons": [
+            "ATG"
+        ],
+        "molecular_weight": 149.21,
+        "name": "Methionine",
+        "short_name": "Met",
+        "side_chain": "Nonpolar"
+    },
+    {
+        "abbreviation": "F",
+        "codons": [
+            "TTT",
+            "TTC"
+        ],
+        "molecular_weight": 165.19,
+        "name": "Phenylalanine",
+        "short_name": "Phe",
+        "side_chain": "Nonpolar"
+    },
+    {
+        "abbreviation": "P",
+        "codons": [
+            "CCT",
+            "CCC",
+            "CCA",
+            "CCG"
+        ],
+        "molecular_weight": 115.13,
+        "name": "Proline",
+        "short_name": "Pro",
+        "side_chain": "Nonpolar"
+    },
+    {
+        "abbreviation": "S",
+        "codons": [
+            "TCT",
+            "TCC",
+            "TCA",
+            "TCG",
+            "AGT",
+            "AGC"
+        ],
+        "molecular_weight": 105.09,
+        "name": "Serine",
+        "short_name": "Ser",
+        "side_chain": "Polar"
+    },
+    {
+        "abbreviation": "T",
+        "codons": [
+            "ACT",
+            "ACC",
+            "ACA",
+            "ACG"
+        ],
+        "molecular_weight": 119.12,
+        "name": "Threonine",
+        "short_name": "Thr",
+        "side_chain": "Polar"
+    },
+    {
+        "abbreviation": "W",
+        "codons": [
+            "TGG"
+        ],
+        "molecular_weight": 204.23,
+        "name": "Tryptophan",
+        "short_name": "Trp",
+        "side_chain": "Nonpolar"
+    },
+    {
+        "abbreviation": "Y",
+        "codons": [
+            "TAT",
+            "TAC"
+        ],
+        "molecular_weight": 181.19,
+        "name": "Tyrosine",
+        "short_name": "Tyr",
+        "side_chain": "Polar"
+    },
+    {
+        "abbreviation": "V",
+        "codons": [
+            "GTT",
+            "GTC",
+            "GTA",
+            "GTG"
+        ],
+        "molecular_weight": 117.15,
+        "name": "Valine",
+        "short_name": "Val",
+        "side_chain": "Nonpolar"
+    }
+]
