+++ src/aaprop_server/main.rs
@@ -1,0 +1,8 @@
+// SPDX-FileCopyrightText: 2023 - 2024 Ali Sajid Imami
+//
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-License-Identifier: MIT
+
+fn main() {
+    println!("This is a server.")
+}
